mosaic/requirements.txt
# platform: linux-64

absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
albumentations==1.2.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.1.1
astroid==2.11.7
asttokens==2.0.8
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==22.1.0
autopep8==1.5.7
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0
bleach==5.0.1
blis==0.7.4
boto3==1.24.57
botocore==1.27.57
cachetools==5.2.0
catalogue==2.0.6
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==1.6.0
cookiecutter==1.7.3
cycler==0.11.0
cymem==2.0.5
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.5.1
docker-pycreds==0.4.0
docutils==0.17.1
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
flake8==4.0.1
flake8-docstrings==1.6.0
fonttools==4.36.0
frozenlist==1.3.1
fsspec==2022.7.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
gmplot==1.4.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.0
idna==3.3
imageio==2.9.0
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.14.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
lazy-object-proxy==1.7.1
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.6.1
mistune==2.0.4
motmetrics==1.2.5
mpmath==1.2.1
multidict==6.0.2
murmurhash==1.0.5
mypy==0.971
mypy-extensions==0.4.3
nanoid==2.0.0
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.6
notebook==6.4.12
numpy==1.23.2
numpydoc==1.1.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pip==22.2.2
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
pluggy==1.0.0
plyfile==0.7.4
poyo==0.5.0
preshed==3.0.5
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pyDeprecate==0.3.1
pydocstyle==6.1.1
pyflakes==2.4.0
Pygments==2.13.0
pylint==2.14.5
pyls-spyder==0.4.0
pyparsing==3.0.9
PyQt5==5.12.3
PyQt5-sip==12.9.0
PyQtWebEngine==5.12.1
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.2
python-slugify==5.0.2
pytorch-lightning==1.5.10
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==5.4.1
pyzmq==23.2.1
QDarkStyle==3.0.2
qstylizer==0.2.1
QtAwesome==1.0.3
qtconsole==5.3.1
QtPy==2.2.0
qudida==0.0.4
regex==2021.8.28
requests==2.28.1
requests-oauthlib==1.3.1
rope==0.19.0
rsa==4.9
Rtree==0.9.7
s3transfer==0.6.0
scalabel==0.3.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.6.1
seaborn==0.11.2
Send2Trash==1.8.0
sentry-sdk==1.9.5
setproctitle==1.3.2
setuptools==65.2.0
setuptools-scm==6.4.2
Shapely==1.8.4
shortuuid==1.0.9
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spacy==3.1.2
spacy-legacy==3.0.8
Sphinx==5.1.1
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.1.3
spyder-kernels==2.1.1
srsly==2.4.1
stack-data==0.4.0
sympy==1.8
tabulate==0.8.10
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
thinc==8.0.9
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
three-merge==0.1.1
tifffile==2022.8.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.4
torch==1.7.1
torchmetrics==0.9.3
torchvision==0.8.2
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typer==0.3.2
typing_extensions==4.3.0
ujson==4.1.0
urllib3==1.26.12
wandb==0.12.21
wasabi==0.8.2
watchdog==2.1.5
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wheel==0.37.1
widgetsnbextension==4.0.2
wrapt==1.14.1
wurlitzer==3.0.2
xmltodict==0.13.0
yapf==0.31.0
yarl==1.8.1
zipp==3.8.1
